+++ backend/canisters/buyback_burn/impl/src/jobs/burn_tokens.rs
@@ -1,101 +1,107 @@
 use crate::state::{ mutate_state, read_state };
 use canister_time::run_now_then_interval;
 use canister_tracing_macros::trace;
-use icrc_ledger_types::icrc1::account::{ Account, Subaccount };
+use icrc_ledger_types::icrc1::account::Account;
 use icrc_ledger_types::icrc1::transfer::TransferArg;
-use std::time::Duration;
-use tracing::debug;
-use tracing::error;
-use tracing::info;
-use types::TokenInfo;
-use crate::utils::fetch_balance_of_sub_account;
-use ic_ledger_types::Tokens;
-
-pub const BUYBACK_BURN_SUB_ACCOUNT: Subaccount = [
-    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-];
-
-const INTERVAL: Duration = Duration::from_secs(7 * 24 * 60 * 60); // 1 week
+use tracing::{ error, info };
+use crate::utils::calculate_burn_amount;
+use candid::Principal;
+use candid::Nat;
 
 pub fn start_job() {
-    run_now_then_interval(INTERVAL, run);
+    let burn_interval = read_state(|s| s.data.burn_config.burn_interval);
+    run_now_then_interval(burn_interval, run);
 }
 
-pub fn run() {}
-
-pub const SUB_ACCOUNT: Subaccount = [
-    1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-];
+// FIXME Process retry here
+pub fn run() {
+    ic_cdk::spawn(process_token_burn());
+}
 
 #[trace]
 pub async fn process_token_burn() {
-    // TODO: Delete extra clone here
     let burn_config = read_state(|s| s.data.burn_config.clone());
-
-    // Get the minting account in order to burn tokens (sending the to minting account is equivalent to burning)
-    let minting_account = Account {
-        owner: burn_config.burn_address,
-        subaccount: None,
-    };
 
     let gldgov_ledger_canister_id = read_state(|s| s.data.gldgov_ledger_canister_id);
 
-    // Specify GLDGov token details
-    let gldgov_token_info = TokenInfo {
-        ledger_id: gldgov_ledger_canister_id,
-        fee: 100_000u64,
-        decimals: 8u64,
-    };
+    // Obtain the canister's GLDGov balance
+    match get_token_balance(gldgov_ledger_canister_id).await {
+        Ok(available_amount) => {
+            // Safely calculate the amount to be burned
+            let amount_to_burn = calculate_burn_amount(available_amount, burn_config.burn_rate);
 
-    // check the reserve pool has enough GLDGov to correctly transfer ( burn )
-    match fetch_balance_of_sub_account(gldgov_token_info.ledger_id, SUB_ACCOUNT).await {
-        Ok(balance) => {
-            // TODO fix here in future. There should be no problem with converting u64 into u128.
-            let total_to_burn =
-                burn_config.min_icp_burn_amount + Tokens::from_e8s(gldgov_token_info.fee);
-            if balance < total_to_burn.e8s() {
-                debug!(
-                    "Balance of reserve pool : {} is too low to make a burn of {} plus a fee of {} ",
-                    balance,
-                    burn_config.min_icp_burn_amount,
-                    gldgov_token_info.fee
-                );
-                return;
+            let min_icp_burn_amount: u128 = burn_config.min_icp_burn_amount.e8s().into();
+
+            // Check if the amount to burn is above the minimum threshold
+            if amount_to_burn < min_icp_burn_amount {
+                match
+                    burn_tokens(
+                        gldgov_ledger_canister_id,
+                        burn_config.burn_address.into(),
+                        amount_to_burn
+                    ).await
+                {
+                    Ok(_) => {
+                        info!("SUCCESS: {} GLDGov tokens burned from the reserve pool.", amount_to_burn);
+                        mutate_state(
+                            |s| {
+                                // FIXME: update the last burn timestamp or other state changes
+                                // s.data.last_burn_time = Some(current_time_ms());
+                            }
+                        );
+                    }
+                    Err(e) => {
+                        error!(
+                            "ERROR: Failed to transfer GLDGov tokens from the reserve pool to the minting account: {:?}",
+                            e
+                        );
+                    }
+                }
             }
         }
         Err(e) => {
-            error!(e);
+            let error_message = format!(
+                "Failed to fetch token balance of the current canister {} from ledger canister ID {}: {:?}",
+                ic_cdk::api::id(),
+                gldgov_ledger_canister_id,
+                e
+            );
+            error!("{}", error_message);
             return;
         }
     }
+}
 
-    // Send tokens to the minting account in order to burn them
+async fn get_token_balance(ledger_id: Principal) -> Result<Nat, String> {
+    icrc_ledger_canister_c2c_client
+        ::icrc1_balance_of(
+            ledger_id,
+            &(Account {
+                owner: ic_cdk::api::id(),
+                subaccount: None,
+            })
+        ).await
+        .map_err(|e| format!("Failed to fetch token balance: {:?}", e))
+}
+
+async fn burn_tokens(
+    ledger_id: Principal,
+    burn_address: Account,
+    amount: u128
+) -> Result<(), String> {
     let args = TransferArg {
-        from_subaccount: Some(BUYBACK_BURN_SUB_ACCOUNT),
-        to: minting_account,
+        from_subaccount: None,
+        to: burn_address,
         fee: None,
         created_at_time: None,
-        amount: burn_config.min_icp_burn_amount.e8s().into(),
+        amount: amount.into(),
         memo: None,
     };
 
-    match icrc_ledger_canister_c2c_client::icrc1_transfer(burn_config.burn_address, &args).await {
-        Ok(_) => {
-            info!(
-                "SUCCESS : {:?} GLDGov tokens burned from reserve pool",
-                burn_config.min_icp_burn_amount
-            );
-            mutate_state(
-                |s| {
-                    // s.data.last_daily_gldgov_burn = Some(current_time_ms);
-                }
-            )
-        }
-        Err(e) => {
-            error!(
-                "ERROR : GLDGov failed to transfer from reserve pool to GLDGov minting account with error : {:?}",
-                e
-            );
-        }
+    // Match the result of the transfer operation
+    match icrc_ledger_canister_c2c_client::icrc1_transfer(ledger_id, &args).await {
+        Ok(Ok(_)) => Ok(()), // Successful transfer
+        Ok(Err(transfer_error)) => Err(format!("Failed to transfer tokens: {:?}", transfer_error)), // Transfer error
+        Err(e) => Err(format!("Failed to send transfer request: {:?}", e)), // Communication or other errors
     }
 }