+++ backend/canisters/buyback_burn/impl/src/jobs/mod.rs
@@ -3,4 +3,5 @@ pub mod swap_tokens;
 
 pub(crate) fn start() {
     burn_tokens::start_job();
+    swap_tokens::start_job();
 }