+++ backend/canisters/buyback_burn/impl/src/token_swap/mod.rs
@@ -2,3 +2,9 @@ pub mod icpswap;
 
 pub mod swap_client;
 pub use swap_client::*;
+
+pub mod swap_clients;
+pub use swap_clients::*;
+
+pub mod swap_config;
+pub use swap_config::*;