+++ backend/canisters/buyback_burn/CHANGELOG.md
@@ -12,4 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 #### Description
 This marks the initial release of Buyback and Burn Canister (buyback_burn_canister).
 
+- TODO: SwapConfig Trait
+- TODO: Job Scheduling
+- TODO: iterate over swap clients
+- TODO: input amount while swap fix it
+
+
 #### Added